refactor(workstation): extract chrome + overlays + dispatchers from inkRuntime (#890 phase 5a.7)
Seventh batch of extractions, building on #902 (5a.6). All chrome
renderers, overlay panels, and the two top-level dispatchers move to
src/workstation/runtime/. After this PR the only render code left in
inkRuntime is what LogInkApp does inline — phase 5b will extract
LogInkApp itself.

6 new files in workstation/runtime/:

  header.ts         (105 LOC) renderHeader — title bar with mode badge,
                              PR glyph, breadcrumb, loading hints
  sidebar.ts        (301 LOC) renderSidebar + renderActiveSidebarContent
                              + renderSelectableSidebarRows
                              + renderActiveStatusTabContent
                              (accordion tabs for status / branches /
                              tags / stashes / worktrees)
  footer.ts         (70 LOC)  renderFooter — contextual + global hints
                              with idle-tip slot
  mainPanel.ts      (137 LOC) renderMainPanel — flat dispatcher to all
                              12 surface renderers + conflicts
  detailPanel.ts    (127 LOC) renderDetailPanel — dispatcher between
                              overlays and detail-family renderers
                              (mostly delegates to surfaces/detail and
                              runtime/overlays)
  overlays.ts       (335 LOC) renderInputPromptPanel,
                              renderConfirmationPanel,
                              renderOnboardingOverlay,
                              renderChordOverlay,
                              renderHelpPanel,
                              renderCommandPalette

inkRuntime.ts: 3,391 -> 2,439 LOC (-952 this PR; -3,600 cumulative
since 5a started, -60% off the original 6,039).

What's left at 2,439 LOC: mostly the LogInkApp React component
(~1,940 LOC) plus orchestration helpers (loadLogInkContext,
loadInkRuntime, safe, startInkInteractiveLog, etc.) and imports.
Phase 5b extracts LogInkApp itself, after which inkRuntime should
land in the 150-300 LOC range.

Big import cleanup: ~25 imports pruned from inkRuntime that became
unused after the dispatchers + chrome + overlays moved out:
  isLogInkContextKeyLoading, isLogInkContextLoading,
  formatBindingKeys, formatLogInkBreadcrumb,
  filterLogInkPaletteCommands, getLogInkChordContinuations,
  getLogInkPaletteCommands, getLogInkFooterHints,
  getLogInkHelpSections, getSidebarVisibleWindow, LogInkTheme,
  branchRowMarker, getPullRequestStateGlyph, sidebarTabCount,
  LogInkSidebarTab, getLogInkSidebarTabs, getLogInkWorkflowActionById,
  LogInkComponents, focusBorderColor, panelTitle, sidebarTabLabel,
  truncate alias, all 12 surface render imports, and the detail
  family imports.

inkRuntime.ts has now lost almost all chrome / surface dependencies —
the remaining imports are exactly what LogInkApp needs: input dispatch,
state model, lifecycle hooks, persistence, refresh watcher. Exactly
the dependency direction the workstation architecture is supposed to
enforce.

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.6). No behavior change — every
chrome panel and overlay renders byte-identically before/after.
