+++ src/commands/log/inkRuntime.ts
@@ -56,19 +56,8 @@ import {
     LogInkContextKey,
     LogInkContextStatus,
     createLogInkContextStatus,
-    isLogInkContextKeyLoading,
-    isLogInkContextLoading,
     updateLogInkContextStatus,
 } from '../../workstation/chrome/context'
-import {
-    formatBindingKeys,
-    formatLogInkBreadcrumb,
-    filterLogInkPaletteCommands,
-    getLogInkChordContinuations,
-    getLogInkPaletteCommands,
-    getLogInkFooterHints,
-    getLogInkHelpSections,
-} from './inkKeymap'
 import {
     LogInkInputKey,
     getInspectorActionsForState,
@@ -77,7 +66,6 @@ import {
 import { hasSeenOnboarding, markOnboardingSeen } from '../../workstation/chrome/onboarding'
 import { getSavedDiffViewMode, saveDiffViewMode } from '../../workstation/chrome/diffViewModePersistence'
 import { getSavedSidebarTab, saveSidebarTab } from '../../workstation/chrome/sidebarPersistence'
-import { getSidebarVisibleWindow } from '../../workstation/chrome/sidebarSelection'
 import {
     PromotedSelectionsSnapshot,
     rectifyPromotedSelectionIndex,
@@ -94,25 +82,14 @@ import {
     LOG_INK_MIN_ROWS,
     getLogInkLayout,
 } from '../../workstation/chrome/layout'
-import { createLogInkTheme, LogInkTheme, LogInkThemeConfig } from '../../workstation/chrome/theme'
-import {
-    branchRowMarker,
-    getPullRequestStateGlyph,
-    sidebarTabCount,
-} from '../../workstation/chrome/iconography'
+import { createLogInkTheme, LogInkThemeConfig } from '../../workstation/chrome/theme'
+import { sortBranches, sortTags } from '../../workstation/chrome/sorting'
 import { IDLE_TIPS_GRACE_MS, IDLE_TIPS_INTERVAL_MS, pickIdleTip } from '../../workstation/chrome/idleTips'
 import {
-    sortBranches,
-    sortTags,
-} from '../../workstation/chrome/sorting'
-import { truncateCells } from '../../workstation/chrome/text'
-import {
-    LogInkSidebarTab,
     LogInkState,
     LogInkView,
     applyLogInkAction,
     createLogInkState,
-    getLogInkSidebarTabs,
     getSelectedInkCommit,
 } from './inkViewModel'
 import { startInteractiveLog } from './interactive'
@@ -187,9 +164,6 @@ import { bisectBad, bisectGood, bisectReset, bisectSkip, extractBisectRemainingH
 import { getCompareDiff } from '../../git/compareData'
 import { getReflogOverview } from '../../git/reflogData'
 import { getTagOverview } from '../../git/tagData'
-import {
-    getLogInkWorkflowActionById,
-} from './inkWorkflows'
 import { getWorktreeListOverview } from '../../git/worktreeData'
 import { WorktreeFileDiff, getWorktreeFileDiff } from '../../git/worktreeDiffData'
 import { canStartLogInkTui, getLogInkRenderOptions } from '../../workstation/chrome/terminal'
@@ -233,52 +207,24 @@ type LogInkOptions = {
 // references continue to work without rewriting every render helper.
 import type {
   LogInkComponentDeps,
-  LogInkComponents,
   LogInkContext,
   LogInkRuntime,
 } from '../../workstation/runtime/types'
-
-// Pure visual helpers shared with future per-surface modules. Moved out of
-// this file so a surface module can render with the same color / glyph rules
-// without pulling in the orchestration code.
-import {
-  focusBorderColor,
-  panelTitle,
-  sidebarTabLabel,
-} from '../../workstation/runtime/utils'
 
 // Promoted-list filter helpers shared by every promoted surface. Live in
 // runtime/ rather than chrome/ because they're tightly coupled to the
 // LogInkState filter-mode shape.
 import { matchesPromotedFilter } from '../../workstation/runtime/promotedFilter'
 
-// Per-surface renderers extracted in phases 5a.1-5a.6 (#890). All 12 main
-// surfaces (+ conflicts + the detail family) now live in their own modules.
-// The chrome render helpers (header, sidebar, footer, overlays, palette)
-// will follow in phase 5a.7.
-import { renderBisectSurface } from '../../workstation/surfaces/bisect'
-import { renderBranchesSurface } from '../../workstation/surfaces/branches'
-import { renderComposeSurface } from '../../workstation/surfaces/compose'
-import { renderConflictsSurface } from '../../workstation/surfaces/conflicts'
-import { renderDiffSurface } from '../../workstation/surfaces/diff'
-import { renderHistoryPanel } from '../../workstation/surfaces/history'
-import { renderPullRequestSurface } from '../../workstation/surfaces/pullRequest'
-import { renderReflogSurface } from '../../workstation/surfaces/reflog'
-import { renderStashSurface } from '../../workstation/surfaces/stash'
-import { renderStatusSurface } from '../../workstation/surfaces/status'
-import { renderTagsSurface } from '../../workstation/surfaces/tags'
-import { renderWorktreesSurface } from '../../workstation/surfaces/worktrees'
-import {
-  renderBranchPreviewPanel,
-  renderCommitDiffDetail,
-  renderCommitPanel,
-  renderComposeContextPanel,
-  renderHistoryInspector,
-  renderStashPreviewPanel,
-  renderTagPreviewPanel,
-} from '../../workstation/surfaces/detail'
-
-const truncate = truncateCells
+// Chrome + overlay + dispatcher renderers extracted in phase 5a.7. The
+// per-surface and detail renderers are consumed internally by mainPanel /
+// detailPanel; LogInkApp just calls these top-level pieces.
+import { renderFooter } from '../../workstation/runtime/footer'
+import { renderHeader } from '../../workstation/runtime/header'
+import { renderSidebar } from '../../workstation/runtime/sidebar'
+import { renderMainPanel } from '../../workstation/runtime/mainPanel'
+import { renderDetailPanel } from '../../workstation/runtime/detailPanel'
+import { renderOnboardingOverlay } from '../../workstation/runtime/overlays'
 
 async function safe<T>(promise: Promise<T>): Promise<T | undefined> {
   try {
@@ -2491,901 +2437,3 @@ function LogInkApp(deps: LogInkComponentDeps): ReactTypes.ReactElement {
   )
 }
 
-function renderHeader(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  columns: number,
-  theme: LogInkTheme,
-  appLabel: string
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const branch = context.branches?.currentBranch || context.provider?.currentBranch || '<detached>'
-  // #784 — surface bisect-in-progress in the title bar so users entering
-  // the TUI mid-bisect see it immediately, before they navigate to gB.
-  const dirtyBase = context.branches?.dirty ? 'dirty' : 'clean'
-  const dirty = context.bisect?.active ? `${dirtyBase} · BISECTING` : dirtyBase
-  const repo = context.provider?.repository.owner && context.provider.repository.name
-    ? `${context.provider.repository.owner}/${context.provider.repository.name}`
-    : 'local repository'
-  const prInfo = context.provider?.currentPullRequest || context.pullRequest?.currentPullRequest
-  const prGlyph = prInfo ? getPullRequestStateGlyph(prInfo, theme) : null
-  const prLabel = prInfo
-    ? `PR #${prInfo.number} ${prInfo.isDraft ? 'DRAFT' : prInfo.state}`
-    : 'no PR'
-  const search = state.filterMode ? `search: ${state.filter}_` : state.filter ? `filter: ${state.filter}` : ''
-  // Boot loading wins over the per-context loading hint because it
-  // tells the user the headline thing they care about (commits aren't
-  // ready yet) — the context fetches finish independently and surface
-  // their own per-section loading copy in the sidebars.
-  const loading = state.bootLoading
-    ? '  loading commits'
-    : isLogInkContextLoading(contextStatus) ? '  loading context' : ''
-  const breadcrumb = formatLogInkBreadcrumb(state.viewStack)
-  const view = breadcrumb ? `  ${breadcrumb}` : ''
-  // Mode indicator (P2.2) — surfaces the current input mode so users
-  // never wonder why `q` doesn't quit while they're editing or filtering.
-  const mode = state.commitCompose.editing
-    ? '[EDIT]'
-    : state.filterMode
-      ? '[FILTER]'
-      : '[NORMAL]'
-  const titlePrefix = `${appLabel}  ${repo}  ${branch}  ${dirty}  `
-  const glyphPart = prGlyph?.glyph ? `${prGlyph.glyph} ` : ''
-  const titleSuffix = `${view}${loading}`
-  const fullTitle = `${titlePrefix}${glyphPart}${prLabel}${titleSuffix}`
-  const titleBudget = columns - mode.length - 4
-  const truncatedTitle = truncate(fullTitle, titleBudget)
-  // Only split into colored fragments when the prefix + glyph + label all
-  // fit unmodified — otherwise the truncate ellipsis can land mid-fragment
-  // and we'd render half a glyph in the wrong color.
-  const splitFragments = truncatedTitle === fullTitle && glyphPart.length > 0
-  const modeColor = theme.noColor
-    ? undefined
-    : state.filterMode || state.commitCompose.editing
-      ? theme.colors.warning
-      : theme.colors.accent
-
-  return h(Box, {
-    borderColor: theme.colors.border,
-    borderStyle: theme.borderStyle,
-    height: 3,
-    paddingX: 1,
-  },
-  splitFragments
-    ? h(Text, { bold: true, color: theme.colors.accent }, titlePrefix)
-    : h(Text, { bold: true, color: theme.colors.accent }, truncatedTitle),
-  splitFragments
-    ? h(Text, { bold: true, color: prGlyph?.color, dimColor: prGlyph?.dim }, glyphPart)
-    : undefined,
-  splitFragments
-    ? h(Text, { bold: true, color: theme.colors.accent }, `${prLabel}${titleSuffix}`)
-    : undefined,
-  h(Text, { bold: true, color: modeColor }, `  ${mode}`),
-  search ? h(Text, { dimColor: true }, `  ${truncate(search, 36)}`) : undefined)
-}
-
-function renderSidebar(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  width: number,
-  bodyRows: number,
-  theme: LogInkTheme
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const focused = state.focus === 'sidebar'
-  const tabs = getLogInkSidebarTabs()
-
-  // Accordion layout — every tab's title is visible on its own line, but
-  // only the active tab expands its content underneath. Switching tabs
-  // (1-5 / [/]) collapses the previous and expands the next.
-  // When sidebar focus has been promoted to the tab header (#806
-  // follow-up), the active tab's title row gets selection styling
-  // and the items below it render without their cursor highlight
-  // (which now lives on the header).
-  const headerFocused = focused && state.sidebarHeaderFocused
-  const tabBlocks = tabs.flatMap((tab, tabIndex) => {
-    const isActive = tab === state.sidebarTab
-    const count = sidebarTabCount(tab, context)
-    const labelWithCount = count !== undefined
-      ? `${sidebarTabLabel(tab)} (${count})`
-      : sidebarTabLabel(tab)
-    const headerText = isActive ? `[${labelWithCount}]` : labelWithCount
-    const headerSelected = isActive && headerFocused
-    const blocks: ReactTypes.ReactElement[] = []
-    if (tabIndex > 0) {
-      blocks.push(h(Text, { key: `tab-spacer-${tab}` }, ''))
-    }
-    blocks.push(h(Text, {
-      key: `tab-header-${tab}`,
-      bold: isActive,
-      dimColor: !isActive,
-      // Selection styling on the header itself when the cursor has
-      // been promoted off the items list. inverse swaps fg/bg so the
-      // highlight reads as "this is the cursor target" identically
-      // to how items render when focused.
-      backgroundColor: headerSelected && !theme.noColor ? theme.colors.selection : undefined,
-      inverse: headerSelected,
-    }, headerText))
-    if (isActive) {
-      blocks.push(...renderActiveSidebarContent(h, Text, tab, state, context, contextStatus, width, bodyRows, theme))
-    }
-    return blocks
-  })
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  },
-  h(Text, { bold: true }, panelTitle('Repository', focused)),
-  h(Text, undefined, ''),
-  ...tabBlocks)
-}
-
-/**
- * Render the indented body of the active sidebar tab. The status tab
- * colours its summary counts (warning / danger / muted) and per-file
- * rows so they read as the same severity scale used in the main status
- * surface; every other tab falls through to `sidebarLines` for its
- * string-based summary.
- */
-function renderActiveSidebarContent(
-  h: typeof ReactTypes.createElement,
-  Text: LogInkComponents['Text'],
-  tab: LogInkSidebarTab,
-  state: LogInkState,
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  width: number,
-  bodyRows: number,
-  theme: LogInkTheme
-): ReactTypes.ReactElement[] {
-  // Available rows for the active tab's list. The sidebar chrome
-  // takes ~10 rows (panel title + spacer + 5 tab headers + 4 inter-tab
-  // spacers); the branches tab eats 3 more for its summary header
-  // (Current / Worktree / spacer). Floor of 8 keeps short terminals
-  // usable; tall terminals (40+ rows) get noticeably more items.
-  const sidebarChrome = 10
-  const branchHeaderRows = tab === 'branches' ? 3 : 0
-  const visibleListCount = Math.max(8, bodyRows - sidebarChrome - branchHeaderRows)
-  if (tab === 'status') {
-    return renderActiveStatusTabContent(h, Text, context, contextStatus, width, theme)
-  }
-
-  // Branches / tags / stashes / worktrees: render selectable rows so
-  // ↑/↓ navigates within the sidebar list and Enter / per-entity keys
-  // act on the cursored item without needing to drill into the
-  // dedicated view (#791 follow-up — in-sidebar selection).
-  // Items render with the cursor highlight only when the sidebar is
-  // focused on this tab AND the cursor is on items (not promoted to
-  // the tab header). The header-focused branch up in `renderSidebar`
-  // owns the highlight in that case.
-  const focused = state.focus === 'sidebar' && state.sidebarTab === tab && !state.sidebarHeaderFocused
-
-  if (tab === 'branches') {
-    if (isLogInkContextKeyLoading(contextStatus, 'branches')) {
-      return [h(Text, { key: 'tab-branches-loading', dimColor: true }, '  Loading branches…')]
-    }
-    const branches = context.branches
-    if (!branches) {
-      return [h(Text, { key: 'tab-branches-empty', dimColor: true }, '  Branches unavailable')]
-    }
-    const sortedBranches = sortBranches(branches.localBranches, state.branchSort)
-    const headerRows: ReactTypes.ReactElement[] = [
-      h(Text, { key: 'tab-branches-current', dimColor: true },
-        truncate(`  Current: ${branches.currentBranch || '<detached>'}`, width - 4)),
-      h(Text, { key: 'tab-branches-state', dimColor: true },
-        `  Worktree: ${branches.dirty ? 'dirty' : 'clean'}`),
-      h(Text, { key: 'tab-branches-spacer' }, ''),
-    ]
-    return [
-      ...headerRows,
-      ...renderSelectableSidebarRows(
-        h, Text, sortedBranches, state.selectedBranchIndex, focused, width, theme,
-        (branch) => `${branchRowMarker(branch, { ascii: theme.ascii })} ${branch.shortName}`,
-        'tab-branches', visibleListCount,
-      ),
-    ]
-  }
-
-  if (tab === 'tags') {
-    if (isLogInkContextKeyLoading(contextStatus, 'tags')) {
-      return [h(Text, { key: 'tab-tags-loading', dimColor: true }, '  Loading tags…')]
-    }
-    const tags = sortTags(context.tags?.tags || [], state.tagSort)
-    if (tags.length === 0) {
-      return [h(Text, { key: 'tab-tags-empty', dimColor: true }, '  No tags found')]
-    }
-    return renderSelectableSidebarRows(
-      h, Text, tags, state.selectedTagIndex, focused, width, theme,
-      (tag) => `${truncate(tag.name, 16)} ${tag.subject}`,
-      'tab-tags', visibleListCount,
-    )
-  }
-
-  if (tab === 'stashes') {
-    if (isLogInkContextKeyLoading(contextStatus, 'stashes')) {
-      return [h(Text, { key: 'tab-stashes-loading', dimColor: true }, '  Loading stashes…')]
-    }
-    const stashes = context.stashes?.stashes || []
-    if (stashes.length === 0) {
-      return [h(Text, { key: 'tab-stashes-empty', dimColor: true }, '  No stashes found')]
-    }
-    return renderSelectableSidebarRows(
-      h, Text, stashes, state.selectedStashIndex, focused, width, theme,
-      (stash, index) => `@{${index}} ${stash.message || '(no message)'}`,
-      'tab-stashes', visibleListCount,
-    )
-  }
-
-  // worktrees
-  if (isLogInkContextKeyLoading(contextStatus, 'worktreeList')) {
-    return [h(Text, { key: 'tab-worktrees-loading', dimColor: true }, '  Loading worktrees…')]
-  }
-  const worktrees = context.worktreeList?.worktrees || []
-  if (worktrees.length === 0) {
-    return [h(Text, { key: 'tab-worktrees-empty', dimColor: true }, '  No linked worktrees')]
-  }
-  return renderSelectableSidebarRows(
-    h, Text, worktrees, state.selectedWorktreeListIndex, focused, width, theme,
-    (worktree) => {
-      const marker = worktree.current ? '*' : ' '
-      const wstate = worktree.dirty ? 'dirty' : 'clean'
-      return `${marker} ${worktree.branch || worktree.path} ${wstate}`
-    },
-    'tab-worktrees', visibleListCount,
-  )
-}
-
-/**
- * Render a sliding-window list of selectable sidebar rows. The cursor
- * highlights the row at `selectedIndex` only when `focused` is true so
- * an unfocused sidebar doesn't compete visually with the active panel.
- * Sliding window keeps the cursor in view as the user navigates a long
- * list; truncation hints surface the count of hidden rows.
- */
-function renderSelectableSidebarRows<T>(
-  h: typeof ReactTypes.createElement,
-  Text: LogInkComponents['Text'],
-  items: T[],
-  selectedIndex: number,
-  focused: boolean,
-  width: number,
-  theme: LogInkTheme,
-  toRowText: (item: T, index: number) => string,
-  keyPrefix: string,
-  visibleCount?: number,
-): ReactTypes.ReactElement[] {
-  if (items.length === 0) return []
-
-  const window = getSidebarVisibleWindow(items.length, selectedIndex, visibleCount)
-  const elements: ReactTypes.ReactElement[] = []
-
-  if (window.truncatedAbove > 0) {
-    elements.push(h(Text, {
-      key: `${keyPrefix}-trunc-above`,
-      dimColor: true,
-    }, truncate(`  … ${window.truncatedAbove} more above`, width - 4)))
-  }
-
-  for (let offset = 0; offset < window.size; offset += 1) {
-    const index = window.start + offset
-    if (index >= items.length) break
-    const isSelected = focused && index === selectedIndex
-    const text = toRowText(items[index], index)
-    elements.push(h(Text, {
-      key: `${keyPrefix}-row-${index}`,
-      backgroundColor: isSelected && !theme.noColor ? theme.colors.selection : undefined,
-      inverse: isSelected,
-    }, truncate(`  ${text}`, width - 4)))
-  }
-
-  if (window.truncatedBelow > 0) {
-    elements.push(h(Text, {
-      key: `${keyPrefix}-trunc-below`,
-      dimColor: true,
-    }, truncate(`  … ${window.truncatedBelow} more below`, width - 4)))
-  }
-
-  return elements
-}
-
-function renderActiveStatusTabContent(
-  h: typeof ReactTypes.createElement,
-  Text: LogInkComponents['Text'],
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  width: number,
-  theme: LogInkTheme
-): ReactTypes.ReactElement[] {
-  if (isLogInkContextKeyLoading(contextStatus, 'worktree')) {
-    return [h(Text, { key: 'tab-status-loading', dimColor: true }, '  Loading status…')]
-  }
-  const worktree = context.worktree
-  if (!worktree) {
-    return [h(Text, { key: 'tab-status-empty', dimColor: true }, '  Status unavailable')]
-  }
-  const colorOf = (state: 'staged' | 'unstaged' | 'untracked'): string | undefined => {
-    if (theme.noColor) return undefined
-    if (state === 'staged') return theme.colors.warning
-    if (state === 'unstaged') return theme.colors.danger
-    return theme.colors.muted
-  }
-  const summaryRow = (count: number, label: string, key: string, kind: 'staged' | 'unstaged' | 'untracked') =>
-    h(Text, { key }, '  ', h(Text, { color: colorOf(kind), bold: count > 0 }, `${count} ${label}`))
-  const fileRows = worktree.files.slice(0, 12).map((file, index) => {
-    const codes = `${file.indexStatus}${file.worktreeStatus}`
-    return h(Text, {
-      key: `tab-status-file-${index}`,
-      color: colorOf(file.state),
-    }, truncate(`  ${codes} ${file.path}`, width - 4))
-  })
-  return [
-    summaryRow(worktree.stagedCount, 'staged', 'tab-status-staged', 'staged'),
-    summaryRow(worktree.unstagedCount, 'unstaged', 'tab-status-unstaged', 'unstaged'),
-    summaryRow(worktree.untrackedCount, 'untracked', 'tab-status-untracked', 'untracked'),
-    ...(fileRows.length
-      ? [h(Text, { key: 'tab-status-spacer' }, ''), ...fileRows]
-      : []),
-  ]
-}
-
-function renderMainPanel(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  worktreeDiff: WorktreeFileDiff | undefined,
-  worktreeDiffLoading: boolean,
-  worktreeHunks: WorktreeHunkOverview | undefined,
-  worktreeHunksLoading: boolean,
-  filePreview: GitCommitFilePreview | undefined,
-  filePreviewLoading: boolean,
-  commitDiffHunkOffsets: number[] | undefined,
-  selectedDetailFile: GitCommitDetail['files'][number] | undefined,
-  stashDiffLines: string[] | undefined,
-  stashDiffLoading: boolean,
-  compareDiffLines: string[] | undefined,
-  compareDiffLoading: boolean,
-  bisectCandidateDetail: GitCommitDetail | undefined,
-  bisectCandidateLoading: boolean,
-  bodyRows: number,
-  width: number,
-  theme: LogInkTheme,
-  hasMoreCommits: boolean,
-  loadingMoreCommits: boolean
-): ReactTypes.ReactElement {
-  if (state.activeView === 'status') {
-    return renderStatusSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'diff') {
-    return renderDiffSurface(
-      h,
-      components,
-      state,
-      context,
-      contextStatus,
-      worktreeDiff,
-      worktreeDiffLoading,
-      worktreeHunks,
-      worktreeHunksLoading,
-      filePreview,
-      filePreviewLoading,
-      commitDiffHunkOffsets,
-      selectedDetailFile,
-      stashDiffLines,
-      stashDiffLoading,
-      compareDiffLines,
-      compareDiffLoading,
-      bodyRows,
-      width,
-      theme
-    )
-  }
-
-  if (state.activeView === 'compose') {
-    return renderComposeSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'branches') {
-    return renderBranchesSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'tags') {
-    return renderTagsSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'reflog') {
-    return renderReflogSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'bisect') {
-    return renderBisectSurface(h, components, state, context, contextStatus, bisectCandidateDetail, bisectCandidateLoading, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'stash') {
-    return renderStashSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'worktrees') {
-    return renderWorktreesSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'pull-request') {
-    return renderPullRequestSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  if (state.activeView === 'conflicts') {
-    return renderConflictsSurface(h, components, state, context, contextStatus, bodyRows, width, theme)
-  }
-
-  return renderHistoryPanel(
-    h,
-    components,
-    state,
-    context,
-    bodyRows,
-    width,
-    theme,
-    hasMoreCommits,
-    loadingMoreCommits
-  )
-}
-
-
-
-
-
-
-
-function renderDetailPanel(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  context: LogInkContext,
-  contextStatus: LogInkContextStatus,
-  detail: GitCommitDetail | undefined,
-  loading: boolean,
-  filePreview: GitCommitFilePreview | undefined,
-  filePreviewLoading: boolean,
-  width: number,
-  tabbed: boolean,
-  theme: LogInkTheme
-): ReactTypes.ReactElement {
-  const focused = state.focus === 'detail'
-
-  if (state.showHelp) {
-    return renderHelpPanel(h, components, state, width, theme, focused)
-  }
-
-  if (state.showCommandPalette) {
-    return renderCommandPalette(h, components, state, width, theme, focused)
-  }
-
-  if (state.inputPrompt) {
-    return renderInputPromptPanel(h, components, state, width, theme, focused)
-  }
-
-  if (state.pendingConfirmationId || state.pendingMutationConfirmation) {
-    return renderConfirmationPanel(h, components, state, width, theme, focused)
-  }
-
-  // which-key style overlay — shows the available chord continuations
-  // when the user has pressed the prefix and we're waiting for the
-  // second key. Mirrors helix / which-key.nvim / doom-emacs.
-  if (state.pendingKey) {
-    return renderChordOverlay(h, components, state, width, theme, focused)
-  }
-
-  // The synthetic "(+) new commit" row routes the inspector through the
-  // worktree summary so the user sees what's staged / unstaged at a glance
-  // — same surface as the compose view's right panel.
-  if (state.activeView === 'history' && state.pendingCommitFocused) {
-    return renderComposeContextPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-
-  // Status + worktree-sourced diff keep the staging compose panel — it's
-  // the action surface for stage / hunk / commit. Commit-sourced diff (from
-  // history → Enter) gets a dedicated explore panel: subject, body, and a
-  // navigable file list whose selection swaps the center diff.
-  if (state.activeView === 'status') {
-    return renderCommitPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-
-  if (state.activeView === 'diff') {
-    if (state.diffSource === 'commit') {
-      return renderCommitDiffDetail(h, components, state, detail, loading, width, theme, focused)
-    }
-    return renderCommitPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-
-  // Compose view: the right panel had been falling through to the inspector
-  // and showing the last selected commit's data, which is wrong context for
-  // an in-progress commit. Show the worktree summary instead.
-  if (state.activeView === 'compose') {
-    return renderComposeContextPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-
-  // Preview pane (P4.1) — fzf / yazi / lazygit style: branches, tags, and
-  // stash views each get a tailored summary of the selected entry instead
-  // of falling through to the (stale) history inspector.
-  if (state.activeView === 'branches') {
-    return renderBranchPreviewPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-  if (state.activeView === 'tags') {
-    return renderTagPreviewPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-  if (state.activeView === 'stash') {
-    return renderStashPreviewPanel(h, components, state, context, contextStatus, width, theme, focused)
-  }
-
-  return renderHistoryInspector(
-    h, components, state, context, contextStatus, detail, loading,
-    filePreview, filePreviewLoading, width, tabbed, theme, focused
-  )
-}
-
-
-function renderInputPromptPanel(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  width: number,
-  theme: LogInkTheme,
-  focused: boolean
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const prompt = state.inputPrompt
-  if (!prompt) {
-    return h(Box, { width })
-  }
-
-  const accent = theme.noColor ? undefined : theme.colors.accent
-  // Multi-line prompts (#806) split on newline and render one Text
-  // row per buffer line — the cursor sits at the end of the last
-  // line via the trailing `_`. Single-line prompts collapse to the
-  // original one-row layout for muscle-memory continuity.
-  const promptLines = prompt.multiline ? prompt.value.split('\n') : [prompt.value]
-  if (promptLines.length === 0) {
-    promptLines.push('')
-  }
-  const valueRows = promptLines.map((line, index) => {
-    const isLast = index === promptLines.length - 1
-    const display = isLast ? `${line}_` : line
-    return h(Text, {
-      key: `prompt-line-${index}`,
-      bold: true,
-      color: accent,
-    }, truncate(display, width - 4))
-  })
-  const hint = prompt.multiline
-    ? 'Enter newline · Ctrl+d submit · Esc cancel · Ctrl+u clear'
-    : 'Enter submit · Esc cancel · Ctrl+u clear'
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  },
-  h(Text, { bold: true }, panelTitle('Prompt', focused)),
-  h(Text, { dimColor: true }, truncate(prompt.label, width - 4)),
-  h(Text, undefined, ''),
-  ...valueRows,
-  h(Text, undefined, ''),
-  h(Text, { dimColor: true }, hint))
-}
-
-function renderConfirmationPanel(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  width: number,
-  theme: LogInkTheme,
-  focused: boolean
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const action = getLogInkWorkflowActionById(state.pendingConfirmationId)
-  const mutationLabel = state.pendingMutationConfirmation === 'revert-hunk'
-    ? 'Revert selected hunk'
-    : state.pendingMutationConfirmation === 'revert-file'
-      ? 'Revert selected file'
-      : state.pendingMutationConfirmation === 'discard-draft'
-        ? 'Quit and discard the in-progress commit draft'
-        : undefined
-  const label = action?.label || mutationLabel || 'Workflow action'
-  const warning = state.pendingMutationConfirmation === 'discard-draft'
-    ? 'You have an unsaved commit draft. Press y to discard it and quit.'
-    : state.pendingMutationConfirmation
-    ? 'This discards local changes and cannot be undone by Coco.'
-    : action?.kind === 'ai'
-    ? `AI action requires confirmation. Estimated ${action.estimatedTokens || '<unknown>'} tokens.`
-    : 'Destructive Git action requires confirmation.'
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  },
-  h(Text, { bold: true }, panelTitle('Confirm', focused)),
-  h(Text, undefined, truncate(label, width - 4)),
-  h(Text, { dimColor: true }, truncate(warning, width - 4)),
-  h(Text, undefined, ''),
-  h(Text, undefined, 'Press y to confirm or n/Esc to cancel.'))
-}
-
-/**
- * First-launch onboarding overlay (P1.3). Shown once per machine, gated
- * by an XDG-style cache marker so subsequent launches go straight to the
- * normal UI. Auto-dismisses on the next keystroke.
- *
- * Replaces the whole layout for the first render rather than overlaying
- * a transient banner — Ink doesn't support floating elements, and a full
- * takeover keeps the message readable on small terminals while still
- * being instantly dismissible.
- */
-function renderOnboardingOverlay(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  rows: number,
-  columns: number,
-  theme: LogInkTheme,
-  appLabel: string
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const accent = theme.noColor ? undefined : theme.colors.accent
-  const tips = [
-    { keys: '?', text: 'open the help panel' },
-    { keys: ':', text: 'open the command palette' },
-    { keys: 'g h', text: 'jump to history (g s status, g d diff, g c compose, g b branches, g t tags, g z stash)' },
-    { keys: '<  esc', text: 'pop the navigation stack / go back' },
-    { keys: '/', text: 'filter the active list' },
-    { keys: 'q  ctrl+c', text: 'quit' },
-  ]
-  const maxKeys = tips.reduce((max, tip) => Math.max(max, tip.keys.length), 0)
-  const lineWidth = Math.max(40, columns - 4)
-
-  return h(Box, {
-    flexDirection: 'column',
-    height: rows,
-    paddingX: 2,
-    paddingY: 1,
-  },
-  h(Text, { bold: true, color: accent }, `Welcome to ${appLabel}`),
-  h(Text, { dimColor: true }, 'A quick keyboard tour — press any key to dismiss.'),
-  h(Text, undefined, ''),
-  ...tips.map((tip, index) => h(Text, { key: `onboarding-tip-${index}` },
-    h(Text, { color: accent, bold: true }, `  ${tip.keys.padEnd(maxKeys)}  `),
-    h(Text, undefined, truncate(tip.text, lineWidth - maxKeys - 4)))),
-  h(Text, undefined, ''),
-  h(Text, { dimColor: true }, 'This tip is shown once per machine. Press any key to continue.'))
-}
-
-/**
- * Which-key style chord overlay (P1.1). When the user presses a chord
- * prefix (currently just `g`), the dispatcher sets `state.pendingKey`
- * and waits for the second key. This panel surfaces the available
- * continuations so newcomers don't have to memorize the chord set.
- *
- * Renders in the detail panel slot; auto-dismisses when the chord
- * completes or `pendingKey` is otherwise cleared.
- */
-function renderChordOverlay(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  width: number,
-  theme: LogInkTheme,
-  focused: boolean
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const prefix = state.pendingKey || ''
-  const continuations = getLogInkChordContinuations(prefix)
-  const accent = theme.noColor ? undefined : theme.colors.accent
-
-  const lines: ReactTypes.ReactNode[] = [
-    h(Text, { key: 'chord-title', bold: true }, panelTitle(`${prefix} … jump`, focused)),
-    h(Text, { key: 'chord-spacer' }, ''),
-  ]
-
-  if (continuations.length === 0) {
-    lines.push(h(Text, {
-      key: 'chord-empty',
-      dimColor: true,
-    }, truncate(`No bindings registered for the ${prefix} prefix.`, width - 4)))
-  } else {
-    for (const entry of continuations) {
-      lines.push(h(Text, { key: `chord-${entry.key}` },
-        h(Text, { color: accent, bold: true }, `  ${entry.key}  `),
-        h(Text, undefined, truncate(`${entry.label.padEnd(10)} ${entry.description}`, width - 9))
-      ))
-    }
-  }
-
-  lines.push(h(Text, { key: 'chord-foot-spacer' }, ''))
-  lines.push(h(Text, {
-    key: 'chord-hint',
-    dimColor: true,
-  }, truncate('press the second key to jump · esc cancels', width - 4)))
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  }, ...lines)
-}
-
-function renderHelpPanel(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  width: number,
-  theme: LogInkTheme,
-  focused: boolean
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const children: ReactTypes.ReactNode[] = [
-    h(Text, { bold: true, key: 'title' }, panelTitle('Help', focused)),
-  ]
-
-  const sections = getLogInkHelpSections({
-    activeView: state.activeView,
-    focus: state.focus,
-  })
-
-  for (const section of sections) {
-    children.push(h(Text, { key: `${section.title}-spacer` }, ''))
-    children.push(h(Text, { bold: true, key: section.title }, section.title))
-    section.bindings.forEach((binding) => {
-      children.push(h(Text, { key: `${section.title}:${binding.id}` },
-        truncate(`${formatBindingKeys(binding).padEnd(14)} ${binding.description}`, width - 4)
-      ))
-    })
-  }
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  }, ...children)
-}
-
-function renderCommandPalette(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  width: number,
-  theme: LogInkTheme,
-  focused: boolean
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  const all = getLogInkPaletteCommands()
-  const filtered = filterLogInkPaletteCommands(all, state.paletteFilter, state.paletteRecent)
-  const recentSet = new Set(state.paletteRecent)
-  const showingRecent = !state.paletteFilter.trim() && state.paletteRecent.length > 0
-
-  const selectedIndex = filtered.length === 0
-    ? 0
-    : Math.max(0, Math.min(state.paletteSelectedIndex, filtered.length - 1))
-
-  // Slide a window of rows around the selection so the cursor stays visible
-  // even with hundreds of bindings.
-  const listRows = 14
-  const startIndex = Math.max(0, selectedIndex - Math.floor(listRows / 2))
-  const visible = filtered.slice(startIndex, startIndex + listRows)
-
-  const inputLine = `> ${state.paletteFilter}_`
-  const matchSummary = filtered.length === 0
-    ? 'no matches'
-    : `${filtered.length} ${filtered.length === 1 ? 'match' : 'matches'}`
-  const hint = '↑/↓ select · enter run · esc close'
-
-  const itemLines = filtered.length === 0
-    ? [h(Text, { key: 'palette-empty', dimColor: true }, 'No commands match the current filter.')]
-    : visible.map((command, offset) => {
-      const index = startIndex + offset
-      const isSelected = index === selectedIndex
-      const cursor = isSelected ? '>' : ' '
-      const recentMarker = showingRecent && recentSet.has(command.id) ? '·' : ' '
-      const kindMarker = command.kind === 'workflow'
-        ? command.workflowKind === 'ai'
-          ? '[AI]'
-          : command.requiresConfirmation
-            ? '[confirm]'
-            : '[action]'
-        : ''
-      const line = `${cursor} ${recentMarker} ${command.keys.padEnd(8)} ${command.label.padEnd(20)} ${kindMarker ? `${kindMarker} ` : ''}${command.description}`
-      return h(Text, {
-        key: `palette-${command.kind}-${command.id}`,
-        bold: isSelected,
-        dimColor: !isSelected,
-      }, truncate(line, width - 4))
-    })
-
-  return h(Box, {
-    borderColor: focusBorderColor(theme, focused),
-    borderStyle: theme.borderStyle,
-    flexDirection: 'column',
-    width,
-    paddingX: 1,
-  },
-  h(Box, { justifyContent: 'space-between' },
-    h(Text, { bold: true }, panelTitle('Command palette', focused)),
-    h(Text, { dimColor: true }, matchSummary)
-  ),
-  h(Text, { color: theme.colors.accent }, truncate(inputLine, width - 4)),
-  h(Text, { dimColor: true }, truncate(hint, width - 4)),
-  h(Text, undefined, ''),
-  ...(showingRecent
-    ? [h(Text, { key: 'palette-recent-hint', dimColor: true }, '· marks recently-used')]
-    : []),
-  ...itemLines)
-}
-
-function renderFooter(
-  h: typeof ReactTypes.createElement,
-  components: LogInkComponents,
-  state: LogInkState,
-  context: LogInkContext,
-  theme: LogInkTheme,
-  idleTip?: string
-): ReactTypes.ReactElement {
-  const { Box, Text } = components
-  // Sidebar item count drives the per-tab footer hints — when items are
-  // present the footer surfaces in-sidebar ops (checkout / apply / pop /
-  // drop), otherwise it falls back to the generic "enter open" hint.
-  const sidebarItemCount = (() => {
-    switch (state.sidebarTab) {
-      case 'branches': return context.branches?.localBranches.length
-      case 'tags': return context.tags?.tags.length
-      case 'stashes': return context.stashes?.stashes.length
-      case 'worktrees': return context.worktreeList?.worktrees.length
-      default: return undefined
-    }
-  })()
-  const hints = getLogInkFooterHints({
-    activeView: state.activeView,
-    diffSource: state.diffSource,
-    diffViewMode: state.diffViewMode,
-    filterMode: state.filterMode,
-    focus: state.focus,
-    pendingKey: state.pendingKey,
-    showCommandPalette: state.showCommandPalette,
-    showHelp: state.showHelp,
-    sidebarTab: state.sidebarTab,
-    sidebarItemCount,
-    compareBaseSet: Boolean(state.compareBase),
-  })
-  // Real status messages always win; idle tips only fill the slot when it
-  // would otherwise be empty.
-  const trailing = state.statusMessage || idleTip || ''
-  const status = trailing ? `  ${trailing}` : ''
-  const contextualText = `${hints.contextual.join('   ')}${status}`
-  const globalText = hints.global.join(' · ')
-
-  return h(Box, {
-    flexDirection: 'row',
-    height: 2,
-    justifyContent: 'space-between',
-    paddingX: 1,
-  },
-  h(Text, { color: theme.colors.muted, dimColor: true }, contextualText),
-  h(Text, { color: theme.colors.muted, dimColor: true }, globalText))
-}
